fix: separate `ActorPackage` and `AIProcess` from `Actor`
