+++ CommonLibSF/include/RE/Starfield.h
@@ -3,9 +3,11 @@
 #include "SFSE/Impl/PCH.h"
 
 #include "RE/A/Actor.h"
+#include "RE/A/ActorPackage.h"
 #include "RE/A/ActorValueInfo.h"
 #include "RE/A/ActorValueOwner.h"
 #include "RE/A/ActorValues.h"
+#include "RE/A/AIProcess.h"
 #include "RE/B/BGSAttackDataForm.h"
 #include "RE/B/BGSDestructibleObjectForm.h"
 #include "RE/B/BGSEditorID.h"
