+++ CommonLibSF/src/RE/A/AIProcess.cpp
@@ -1,0 +1,4 @@
+#include "RE/A/AIProcess.h"
+namespace RE
+{
+}
