+++ CommonLibSF/src/RE/A/ActorPackage.cpp
@@ -1,0 +1,4 @@
+#include "RE/A/ActorPackage.h"
+namespace RE
+{
+}
